refactor: collapse one-caller modules and single-use indirection
Each of these was a file or export whose only job was to forward to the
next one:

- wretchApi and popup/provider/utils had one caller apiece, inlined into it.
- OpenForumLinks was three files (container, button, hook) used once each in
  sequence; now one component.
- web-ext.config.ts existed to set a single flag, now webExt in wxt.config.
- utils/url.ts is replaced by toQueryString, which keeps the widening the
  callers need (interfaces have no implicit index signature, hence `object`
  rather than Record<string, string>). Its deserialize half was a bare
  URLSearchParams wrapper and is inlined at both sites.
- filterRecord keeps its five callers but loses the hand-rolled reduce.
- Four types and one helper were exported while only ever used in their own
  file.
