--- apps/extension/src/entrypoints/popup/panels/HomePopup/components/OpenForumLinks/useFeedbackButton.ts
@@ -1,40 +1,0 @@
-import { noOp } from '@bypass/shared';
-import { useEffect, useState } from 'react';
-
-export enum EButtonState {
-  INITIAL,
-  LOADING,
-  SUCCESS,
-}
-
-const SUCCESS_TIMEOUT_MS = 3000;
-
-const useFeedbackButton = (handler: () => Promise<void>) => {
-  const [buttonState, setButtonState] = useState(EButtonState.INITIAL);
-
-  useEffect(() => {
-    if (buttonState !== EButtonState.SUCCESS) {
-      return noOp;
-    }
-    const timeout = setTimeout(
-      () => setButtonState(EButtonState.INITIAL),
-      SUCCESS_TIMEOUT_MS
-    );
-    return () => clearTimeout(timeout);
-  }, [buttonState]);
-
-  const onClick = async () => {
-    setButtonState(EButtonState.LOADING);
-
-    await handler();
-
-    setButtonState(EButtonState.SUCCESS);
-  };
-
-  return {
-    buttonState,
-    onClick,
-  };
-};
-
-export default useFeedbackButton;
