+++ apps/extension/src/entrypoints/popup/panels/PersonsPanel/utils/uploadImage.ts
@@ -1,9 +1,17 @@
-import { wretchApi } from '@/apis/wretchApi';
+import wretch from 'wretch';
+
+import { env } from '@/constants/env';
+import useFirebaseStore from '@/store/firebase/useFirebaseStore';
 
 export const uploadFileToFirebase = async (blob: Blob, fileName: string) => {
   const formData = new FormData();
   formData.append('file', blob, fileName);
 
-  const wretch = await wretchApi();
-  await wretch.url('/upload-file').body(formData).post().res();
+  const { getIdToken } = useFirebaseStore.getState();
+  await wretch(`${env.NEXT_PUBLIC_HOST_NAME}/api`)
+    .auth(`Bearer ${await getIdToken()}`)
+    .url('/upload-file')
+    .body(formData)
+    .post()
+    .res();
 };
