--- apps/extension/web-ext.config.ts
@@ -1,5 +1,0 @@
-import { defineWebExtConfig } from 'wxt';
-
-export default defineWebExtConfig({
-  disabled: true,
-});
