--- packages/shared/src/utils/url.ts
@@ -1,5 +1,0 @@
-export const serializeObjectToQueryString = (urlParams: Record<string, any>) =>
-  new URLSearchParams(urlParams).toString();
-
-export const deserializeQueryStringToObject = (queryString: string) =>
-  Object.fromEntries(new URLSearchParams(queryString));
